+++ subpackages/cli/source/odood/cli/commands/repository.d
@@ -8,7 +8,7 @@ private import thepath: Path;
 
 private import odood.cli.core: OdoodCommand;
 private import odood.lib.project: Project;
-private import odood.lib.odoo.utils: fixVersionConflict;
+private import odood.lib.odoo.utils: fixVersionConflict, updateManifestSerie;
 
 
 class CommandRepositoryAdd: OdoodCommand {
@@ -88,11 +88,33 @@ class CommandRepositoryFixVersionConflict: OdoodCommand {
 }
 
 
+class CommandRepositoryFixSerie: OdoodCommand {
+    this() {
+        super(
+            "fix-series",
+            "Fix series in manifests of addons in this repo. Set series to project's serie");
+        this.add(new Argument(
+            "path", "Path to repository to fix conflicts in.").optional());
+    }
+
+    public override void execute(ProgramArgs args) {
+        auto project = Project.loadProject;
+
+        auto repo = project.addons.getRepo(
+            args.arg("path") ? Path(args.arg("path")) : Path.current);
+        foreach(addon; repo.addons)
+            addon.path.join("__manifest__.py").updateManifestSerie(
+                    project.odoo.serie);
+    }
+}
+
+
 class CommandRepository: OdoodCommand {
     this() {
         super("repo", "Manage git repositories.");
         this.add(new CommandRepositoryAdd());
         this.add(new CommandRepositoryFixVersionConflict());
+        this.add(new CommandRepositoryFixSerie());
     }
 }